// src/client/loginSchema.ts
import { z } from "zod";

export const loginSchema = z.object({
  params: z.object({
    playerWallSettings: z.number(),
    subscriptionSourceType: z.number(),
    numLogins: z.number(),
    numAJHQGiftCards: z.number(),
    activePetInvId: z.number(),
    uuid: z.string().uuid(),
    webWallStatus: z.number(),
    createdAt: z.number().transform(v => new Date(v * 1000)),
    denPrivacySettings: z.string().transform(v => parseInt(v)),
    recyclePercentage: z.number(),
    bambooCount: z.number(),
    accountType: z.number(),
    jammaDate: z.number().optional(),
    sgChatType: z.string().transform(v => parseInt(v)),
    perUserAvId: z.number(),
    newspaperDefs: z.array(z.object({ defId: z.number(), ts: z.number().transform(v => new Date(v * 1000)) })),
    woodCount: z.number(),
    numAJHQBulkGiftCards: z.number(),
    lastBroadcastMessage: z.string(),
    numDaysLeftOnSubscription: z.number(),
    isGuide: z.boolean(),
    sgChatTypeNonDegraded: z.string().transform(v => parseInt(v)),
    hasOnlineBuddies: z.boolean(),
    gemstoneCount: z.number(),
    isModerator: z.boolean(),
    silverCount: z.number(),
    diamondsCount: z.number(),
    ecoCreditsCount: z.number(),
    avatarName: z.string().transform(s => s.split("|").map(e => parseInt(e))).optional(),
    avName: z.string().transform(s => s.split("|").map(e => parseInt(e))).optional(),
    dbUserId: z.number(),
    orbsCount: z.number(),
    interactions: z.number(),
    numRedemptionCards: z.number(),
    ecoCredits: z.number(),
    accountTypeChanged: z.boolean(),
    goldCount: z.number(),
    email: z.string().email().optional(),
    strawCount: z.number(),
    gemsCount: z.number(),
    activeDenRoomInvId: z.number(),
    pendingFlags: z.number(),
    eCardPrivacySettings: z.number(),
    stoneCount: z.number(),
    sessionId: z.number(),
    userName: z.string(),
    userNameModerated: z.number().transform(v => v !== 0),
    userId: z.number(),
    ticketsCount: z.number(),
    numUnreadECards: z.number(),
    dailyGiftIndex: z.number(),
    usableAdoptAPetDefs: z.array(z.unknown())
  })
})